Code cleanup; removing K&R style C, etc.
Clean up dinit, dinitutl, dtypeutl, and listing.

+++ tools/flang2/flang2exe/dtypeutl.c
@@ -85,7 +85,7 @@ is_array_dtype(DTYPE dtype)
 DTYPE
 array_element_dtype(DTYPE dtype)
 {
-  return is_array_dtype(dtype) ? DTY(dtype + 1) : DT_NONE;
+  return is_array_dtype(dtype) ? DTY((int)dtype + 1) : DT_NONE;
 }
 
 static LOGICAL
